+++ CHANGELOG.md
@@ -8,8 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 This file was automatically generated for [incolumepy.utils](https://gitlab.com/development-incolume/incolumepy.utils/-/tree/2.5.4)
 
 ---
+## [0.4.0]	2022-03-31:
+	Added: Protótipo dataclass para atos acrescentado
 ## [0.3.0]	2022-03-29:
-	[This Is Why Python Data Classes Are Awesome](https://www.youtube.com/watch?v=CvQ7e6yUtnw)
+	Added: [This Is Why Python Data Classes Are Awesome](https://www.youtube.com/watch?v=CvQ7e6yUtnw)
 ## [0.2.1]	2022-03-29:
 	Changed: Fix links from CHANGELOG.md
 ## [0.2.0]	2022-03-29:
@@ -21,3 +23,4 @@ This file was automatically generated for [incolumepy.utils](https://gitlab.com/
 [0.2.0]: https://github.com/incolumepy-prospections/incolumepy.dataclass/compare//0.1.0...0.2.0
 [0.2.1]: https://github.com/incolumepy-prospections/incolumepy.dataclass/compare//0.2.0...0.2.1
 [0.3.0]: https://github.com/incolumepy-prospections/incolumepy.dataclass/compare//0.2.1...0.3.0
+[0.4.0]: https://github.com/incolumepy-prospections/incolumepy.dataclass/compare//0.3.0...0.4.0